Refactor: Separate business logic from routes
* Adds api/ directory to contain route business logic
* Adds a V3Router.ts file for all routes for /api/v3 endpoints
* Adds `makeApiResponse` utility function
* Adds a StoreDocument union type
