+++ src/db.ts
@@ -1,11 +1,9 @@
 import nano from 'nano'
 
 import { config } from './config'
-import { StoreDirectory, StoreFile, StoreRepo } from './types'
+import { StoreDocument } from './types'
 
 const url = `http://admin:${config.couchAdminPassword}@${config.couchHost}:${config.couchPort}`
-
-type AllDocumentTypes = StoreRepo | StoreDirectory | StoreFile
 
 const database =
   typeof process.env.DATABASE === 'string'
@@ -13,6 +11,6 @@ const database =
     : 'sync_datastore'
 
 const connection = nano(url)
-const dataStore = connection.use<AllDocumentTypes>(database)
+const dataStore = connection.use<StoreDocument>(database)
 
 export { connection as nano, dataStore }
