+++ src/routes/config.ts
@@ -1,7 +1,7 @@
 import Router from 'express-promise-router'
 
 import { config } from '../config'
-import { ApiResponse } from '../types'
+import { makeApiResponse } from '../utils'
 
 interface ConfigGetResponseData {
   maxPageSize: number
@@ -11,11 +11,7 @@ export const configRouter = Router()
 
 configRouter.get('/config', async (req, res) => {
   // Send response
-  const response: ApiResponse<ConfigGetResponseData> = {
-    success: true,
-    data: {
-      maxPageSize: config.maxPageSize
-    }
-  }
-  res.status(200).json(response)
+  res.status(200).json(
+    makeApiResponse<ConfigGetResponseData>({ maxPageSize: config.maxPageSize })
+  )
 })
